功能添加

+++ reggie_server/src/main/java/com/reggie/service/impl/CategoryServiceImpl.java
@@ -2,14 +2,17 @@ package com.reggie.service.impl;
 
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
+import com.reggie.constant.MessageConstant;
 import com.reggie.constant.StatusConstant;
 import com.reggie.dto.CategoryDTO;
 import com.reggie.dto.CategoryPageQueryDTO;
 import com.reggie.entity.Category;
+import com.reggie.exception.DeletionNotAllowedException;
 import com.reggie.mapper.CategoryMapper;
+import com.reggie.mapper.DishMapper;
+import com.reggie.mapper.SetmealMapper;
 import com.reggie.result.PageResult;
 import com.reggie.service.CategoryService;
-import io.swagger.v3.oas.annotations.parameters.RequestBody;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,6 +24,12 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Autowired
     private CategoryMapper categoryMapper;
+
+    @Autowired
+    private DishMapper dishMapper;
+
+    @Autowired
+    private SetmealMapper setmealMapper;
 
     @Override
     public void save(CategoryDTO categoryDTO) {
@@ -41,6 +50,14 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void deleteById(Long id) {
+        Long dishCount = dishMapper.countByCategoryId(id);
+        if (dishCount > 0){
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_DISH);
+        }
+        Long setmealCount = setmealMapper.countByCategoryId(id);
+        if (setmealCount > 0){
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_SETMEAL);
+        }
         categoryMapper.deleteBuId(id);
     }
 }
